v1.4.2 strategy buy-market sell-limit

+++ lib/domain/model/trading/model_open_order.dart
@@ -8,15 +8,15 @@ class ModelOpenOrder{
 
 
  String createdAt;
-  int filledSize;
+  double filledSize;
   String future;
   int id;
   String market;
   double price;
  double avgFillPrice;
-  int remainingSize;
+  double remainingSize;
   String side;
-  int size;
+  double size;
   String status;
   String type;
   bool reduceOnly;
